+++ packages/jaspr/test/foundation/styles/groups/groups_test.dart
@@ -67,7 +67,8 @@ void main() {
           letterSpacing: Unit.em(0.5),
           wordSpacing: Unit.points(10),
           lineHeight: Unit.rem(1.5),
-          textShadow: TextShadow(offsetX: Unit.zero, offsetY: Unit.zero, blur: Unit.pixels(10)),
+          textShadow: TextShadow(
+              offsetX: Unit.zero, offsetY: Unit.zero, blur: Unit.pixels(10)),
           textOverflow: TextOverflow.ellipsis,
         );
 
@@ -136,7 +137,8 @@ void main() {
           width: Unit.percent(80),
           height: Unit.percent(70),
           maxWidth: Unit.pixels(1000),
-          border: Border.only(top: BorderSide(style: BorderStyle.solid, color: Colors.green)),
+          border: Border.only(
+              top: BorderSide(style: BorderStyle.solid, color: Colors.green)),
           radius: BorderRadius.circular(Unit.pixels(20)),
           outline: Outline(color: Colors.yellow),
           overflow: Overflow.visible,
@@ -145,7 +147,28 @@ void main() {
           zIndex: ZIndex(100),
           opacity: 0.5,
           transform: Transform.scale(2),
-          shadow: BoxShadow(offsetX: Unit.zero, offsetY: Unit.zero, blur: Unit.pixels(10)),
+          shadow: BoxShadow(
+              offsetX: Unit.zero, offsetY: Unit.zero, blur: Unit.pixels(10)),
+          filter: Filter([
+            FilterFn.blur(Unit.rem(2)),
+            FilterFn.brightness(0.0),
+            FilterFn.contrast(0.1),
+            FilterFn.dropShadow(
+                offsetX: Unit.rem(1),
+                offsetY: Unit.rem(1),
+                spread: Unit.rem(0.2),
+                color: Colors.red),
+            FilterFn.grayscale(0.3),
+            FilterFn.hueRotate(Angle.deg(45))
+          ]),
+          backdropFilter: BackdropFilter([
+            FilterFn.invert(0.4),
+            FilterFn.opacity(0.5),
+            FilterFn.sepia(0.6),
+            FilterFn.saturate(0.7),
+            FilterFn.url("abc.svg#urltest"),
+            FilterFn.src("xyz.svg#srctest"),
+          ]),
           cursor: Cursor.crosshair,
           transition: Transition('width', duration: 500),
         );
@@ -172,6 +195,8 @@ void main() {
             'opacity': '0.5',
             'transform': 'scale(2)',
             'box-shadow': '0 0 10px',
+            'filter': 'blur(2rem) brightness(0) contrast(0.1) drop-shadow(1rem 1rem 0.2rem red) grayscale(0.3) hue-rotate(45deg)',
+            'backdrop-filter': 'invert(0.4) opacity(0.5) sepia(0.6) saturate(0.7) url(abc.svg#urltest) src(xyz.svg#srctest)',
             'cursor': 'crosshair',
             'transition': 'width 500ms',
           }),
@@ -234,7 +259,11 @@ void main() {
           ])),
           gap: Gap(row: Unit.pixels(20)),
           autoRows: [TrackSize(Unit.percent(20)), TrackSize.auto],
-          autoColumns: [TrackSize(Unit.pixels(100)), TrackSize.auto, TrackSize.auto],
+          autoColumns: [
+            TrackSize(Unit.pixels(100)),
+            TrackSize.auto,
+            TrackSize.auto
+          ],
           justifyItems: JustifyItems.center,
         );